Fixed type errors caused by strict

+++ tests/TestCase/Job/Engine/RedisEngineTest.php
@@ -81,7 +81,7 @@ class RedisEngineTest extends OriginTestCase
         Queue::config('redis-test', [
             'engine' => 'Redis',
             'host' => env('REDIS_HOST'),
-            'port' => env('REDIS_PORT'),
+            'port' => (int) env('REDIS_PORT'),
         ]);
 
         $this->engine = Queue::connection('redis-test');
